Migrate the timer and count values from Setting to Category. Checkpoint for running code. Still missing how to edit those two values under Category.
- Migrate the Import/Export/Delete Questions from routes_main.py to routes_settings.py
- Migrate the `timer_duration` and `questions_per_quiz` to Category model and remove Setting model
- Update routes_quiz.py to take into accound change to Category change
- Update routes_settings.py and settings.html for migration
- Fix manage.py app location for flask_migrate to work
- TODO how to edit timer and count values
