+++ modules/routes_settings.py
@@ -1,46 +1,102 @@
 """
 Copyright © 2024 J. Michael McGarrah <mcgarrah@gmail.com>
 """
-from flask import render_template, request, redirect, url_for
-from modules.models import db, Setting
+import json
+from flask import render_template, request, redirect, url_for, make_response
+from modules.models import db, Question
 
 def settings():
     """
-    Retrieves the current settings for `timer_duration` and `num_questions`.
-    Uses default values if settings are not found.
-    Renders the `settings.html` template with these values.
+    Renders the `settings.html` template.
     """
-    timer_setting = Setting.query.filter_by(name='timer_duration').first()
-    num_questions_setting = Setting.query.filter_by(name='num_questions').first()
-    timer_duration = int(timer_setting.value) if timer_setting else 300
-    num_questions = int(num_questions_setting.value) if num_questions_setting else 5
-    return render_template('settings.html',
-                           timer_duration=timer_duration, num_questions=num_questions)
+    return render_template('settings.html')
 
-def update_settings():
+def import_questions(file_path=None):
     """
-    Updates the settings for timer_duration and num_questions based on the form submission.
-    Checks if the settings exist; if not, creates new settings.
-    Commits the changes to the database.
-    Redirects to the settings page.
+    Import questions from a JSON file that is either uploaded or specified by file path, 
+    and add them to the database.
     """
-    timer_duration = request.form['timer_duration']
-    num_questions = request.form['num_questions']
-
-    timer_setting = Setting.query.filter_by(name='timer_duration').first()
-    num_questions_setting = Setting.query.filter_by(name='num_questions').first()
-
-    if timer_setting:
-        timer_setting.value = timer_duration
+    if file_path:
+        with open(file_path, encoding="utf-8") as f:
+            data = json.load(f)
     else:
-        new_timer_setting = Setting(name='timer_duration', value=timer_duration)
-        db.session.add(new_timer_setting)
+        file = request.files.get('file')
+        if not file:
+            with open('initial_questions.json', encoding="utf-8") as f:
+                data = json.load(f)
+        else:
+            data = json.load(file)
 
-    if num_questions_setting:
-        num_questions_setting.value = num_questions
-    else:
-        new_num_questions_setting = Setting(name='num_questions', value=num_questions)
-        db.session.add(new_num_questions_setting)
+    for category in data:
+        category_obj = Category.query.filter_by(name=category['category']).first()
+        if not category_obj:
+            category_obj = Category(name=category['category'])
+            db.session.add(category_obj)
+            db.session.commit()
 
+        for question in category['questions']:
+            question_obj = Question(
+                question=question['question'],
+                options=json.dumps(question['options']),
+                answer=question['answer'],
+                answer_details=question.get('answer_details', ''),
+                no_shuffle=question.get('no_shuffle', False),
+                category_id=category_obj.id
+            )
+            db.session.add(question_obj)
     db.session.commit()
+    return redirect(url_for('settings')) if not file_path else None
+
+def export_questions():
+    """
+    Export the categories and questions from database to a JSON file.
+    """
+    categories = Category.query.all()
+    export_data = []
+
+    for category in categories:
+        questions = Question.query.filter_by(category_id=category.id).all()
+        questions_list = []
+        for question in questions:
+            questions_list.append({
+                'question': question.question,
+                'options': json.loads(question.options),
+                'answer': question.answer,
+                'answer_details': question.answer_details,
+                'no_shuffle': question.no_shuffle
+            })
+
+        export_data.append({
+            'category': category.name,
+            'questions': questions_list
+        })
+
+    # TOOD: Verify below code matches this section of code commented out
+    # with open('exported_questions.json', 'w', encoding="utf-8") as f:
+    #     json.dump(export_data, f, indent=4)
+    # return send_file('exported_questions.json', as_attachment=True)
+
+    response = make_response(json.dump(export_data, f, indent=4))
+    response.headers['Content-Disposition'] = 'attachment; filename=exported_questions.json'
+    response.mimetype = 'application/json'
+    return response
+
+def clear_questions():
+    """
+    Clear all questions from the database and remove unused categories.
+    """
+    # Delete all questions
+    Question.query.delete()
+
+    # Commit changes to remove questions
+    db.session.commit()
+
+    # Find and delete categories with no questions
+    unused_categories = Category.query.outerjoin(Question).filter(Question.id is None).all()
+    for category in unused_categories:
+        db.session.delete(category)
+
+    # Commit changes to remove unused categories
+    db.session.commit()
+
     return redirect(url_for('settings'))
