+++ templates/settings.html
@@ -4,20 +4,7 @@
 
 {% block content %}
 <div class="container mt-4">
-    <h1 class="mt-4">Settings</h1>
-    <form action="/update_settings" method="POST">
-        <div class="form-group">
-            <label for="timer_duration">Quiz Timer Duration (seconds):</label>
-            <input type="number" id="timer_duration" name="timer_duration" required class="form-control" value="{{ timer_duration }}">
-        </div>
-        <div class="form-group mt-3">
-            <label for="num_questions">Number of Questions per Quiz:</label>
-            <input type="number" id="num_questions" name="num_questions" required class="form-control" value="{{ num_questions }}">
-        </div>
-        <button type="submit" class="btn btn-primary mt-3">Update Settings</button>
-    </form>
-
-    <h2 class="mt-4">Manage Questions</h2>
+    <h1 class="mt-4">Manage Questions</h1>
     <form action="/import_questions" method="POST" enctype="multipart/form-data" class="mt-3">
         <div class="form-group">
             <label for="file">Import Questions (JSON file):</label>
